improve baseEntityJsonParser
